Added Javadoc comments for all API members.
Added the file Overview.html, that describes the SDK as a whole.

Added five files called package.html, that describe each package.

Configured the Eclipse project to include links to the Javadocs for the
referenced libraries (the official JDK docs, commons-codec, and gson).
All of these types are now hot-linked on our docs.

Created an Ant build.xml script for building the HTML docs. This script
includes JavaScript for automatically generating information in the
header, footer, and bottom of each web page (date stamp, product name
and version, copyright information.